stages: add `coreos.live-iso` stage
This adds a new `org.osbuild.coreos.live-iso` stage to build a CoreOS
live ISO. The code is heavily based on the `cmd-buildextend-live` script
from coreos-assembler, but a lot of things had to be adapted:
- the stage is provided the deployed oscontainer tree, metal, and
  metal4k images as inputs
- we use chroot instead of supermin to execute some commands in the
  context of the target oscontainer
- we use a loopback device to create `efiboot.img` instead of libguestfs
- a bunch of calls that were wrapped by libguestfs for us (e.g.
  mkfs.vfat, mksquashfs), we now have to call ourselves; to retain
  maximum compatibility, I ensured that we still effectively use the
  same args that libguestfs passed
- we require read-write access to make certain changes, so we use
OverlayFS instead of mounting the filesystem with the rw option.

And various other minor adjustments.

Of course, this is not really in the spirit of the osbuild philosophy
of having smaller-scoped stages. The goal should be to eventually break
this down in many stages that we can then assemble into a pipeline.
However, this will take a lot of time and effort. The CoreOS live ISO
has _many_ special things going on and exposes APIs that are heavily
relied on by both OpenShift and end-users. Having this code live here at
least gives a way to do this incrementally.

It also more quickly unlocks a major benefit: customized CoreOS live
ISOs. The only real top-level input is the oscontainer (the metal
images are derivations from that), and this new stage can be used with
customized CoreOS oscontainers.

Co-authored-by: Dusty Mabe <dusty@dustymabe.com>
Co-authored-by: Renata Ravanelli <renata.ravanelli@gmail.com>

Signed-off-by: Renata Ravanelli <rravanel@redhat.com>